Do not replace regex if .*
if the block regex is .*, do not replace it with other entries.
this comes into play when we are blocking an edge by using multiple
regex.
for eg. for 4.7.4, there are two regex used, `.*` and `4\.6\..*`
In some cases, cincinnati would replace `.*` with `4\.6\..*` which
was causing cincinnati to create an incorrect and inconsistant graph.
this PR adds a condition to check if the existing regex is `.*` and
does not replace it with any other regex.