add multi arch images support
this commit adds support for multi arch images.
we are using ManifestLists to represent multi arch images

As manifest list images only contain digests of the
images contained in the manifest, the layers_digests
function from dkregistry returns the digest of all the images
contained in the ManifestList instead of individual
layers of the manifests.
The layers of a specific image from manifest list is
obtained using the digest of the image from the
manifest list and getting it's manifest and manifestref
(get_manifest_and_ref()) and using this manifest of
the individual image to get the layers.